+++ packages/explorerkit-server/package.json
@@ -32,10 +32,12 @@
     "@types/body-parser": "^1.19.5",
     "@types/express": "^4.17.21",
     "@types/node": "^20.10.4",
+    "@types/supertest": "^2.0.16",
     "@vitest/coverage-v8": "^0.34.2",
     "eslint": "^8.47.0",
     "eslint-config-explorerkit": "workspace:*",
     "nodemon": "^3.0.2",
+    "supertest": "^6.3.3",
     "vitest": "^0.34.2"
   },
   "dependencies": {
